backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
openai==1.10.0
httpx==0.26.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
Pillow==10.2.0
lxml==5.1.0
aiofiles==23.2.1

